image-processing/requirements.txt
# Add these Azure packages
azure-identity>=1.14.0
azure-storage-queue>=12.8.0
azure-storage-blob>=12.19.0
azure-data-tables>=12.4.0
azure-core>=1.29.5

# Existing packages needed
pika
Pillow
ninja
pybind11

diffusers
einops
opencv-python
numpy
torch
transformers
torchvision
#taming-transformers-rom1504
#ConfigArgParse
#ipdb
omegaconf

#sentencepiece
tqdm

# Mesh Processing
trimesh
pymeshlab
pygltflib
xatlas
#kornia
#facexlib

# Training
accelerate
#pytorch_lightning
#scikit-learn
#scikit-image

# Demo only
gradio
fastapi
uvicorn
rembg
onnxruntime
#gevent
#geventhttpclient
